// citpwawcl/src/app/models/dati-sc-model.ts
export class DatiSCModel {
    idTipoComponente: string;
    progressivo: number;
    dataInstall: string;
    codiceImpianto: number;
    dataDismiss: string;
    flgDismissione: number;
    dataUltMod: string;
    utenteUltMod: string;
    fkMarca: number;
    descMarca: string;
    matricola: string;
    modello: string;
    potenzaTermicaKw: number;
    note: string;
    tempoManutAnni: number;
    dataMinimaControllo: string;
    dataMassimaControllo: string;
    nomeProprietario: string;
    cfProprietario: string;
    cf: string;
    siglaRea: string;
    numeroRea: string;
}